unit test working in float

+++ test/backend_test.in.cpp
@@ -9586,8 +9586,8 @@ NGRAPH_TEST(${BACKEND_NAME}, quantize)
     Shape shape{4, 3};
     auto X = make_shared<op::Parameter>(element::f32, shape);
     auto Scale = make_shared<op::Parameter>(element::f32, Shape{1});
-    auto Offset = make_shared<op::Parameter>(element::u8, Shape{1});
-    auto f = make_shared<Function>(make_shared<op::Quantize>(X, Scale, Offset, element::u8, AxisSet{}), op::ParameterVector{X, Scale, Offset});
+    auto Offset = make_shared<op::Parameter>(element::f32, Shape{1});
+    auto f = make_shared<Function>(make_shared<op::Quantize>(X, Scale, Offset, element::f32, AxisSet{}), op::ParameterVector{X, Scale, Offset});
     auto backend = runtime::Backend::create("${BACKEND_NAME}");
 
     auto x = backend->create_tensor(element::f32, shape);
@@ -9596,11 +9596,11 @@ NGRAPH_TEST(${BACKEND_NAME}, quantize)
     auto scale = backend->create_tensor(element::f32, Shape{1});
     copy_data(scale, vector<float>{2});
 
-    auto offset = backend->create_tensor(element::u8, Shape{1});
-    copy_data(offset, vector<uint8_t>{1});
+    auto offset = backend->create_tensor(element::f32, Shape{1});
+    copy_data(offset, vector<float>{1});
 
-    auto result = backend->create_tensor(element::u8, shape);
+    auto result = backend->create_tensor(element::f32, shape);
 
     backend->call_with_validate(f, {result}, {x, scale, offset});
-    EXPECT_EQ((vector<uint8_t>{7, 2, 6, 6, 5, 3, 4, 2, 4, 3, 7, 5 }), read_vector<uint8_t>(result));
+    EXPECT_EQ((vector<float>{7, 2, 6, 6, 5, 3, 4, 2, 4, 3, 7, 5 }), read_vector<float>(result));
 }
